+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
-## [Unreleased]
+## [0.2.0] - 2016-12-13
 ### Added
 - Add missing definitions to index.d.ts
 - Add fallback for incorrect builds/locales
@@ -18,5 +18,6 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 - Code for initial release
 
-[Unreleased]: https://github.com/HearthSim/npm-hearthstonejson/compare/0.1.1...HEAD
+[Unreleased]: https://github.com/HearthSim/npm-hearthstonejson/compare/0.2.0...HEAD
+[0.2.0]: https://github.com/HearthSim/npm-hearthstonejson/compare/0.1.1...0.2.0
 [0.1.1]: https://github.com/HearthSim/npm-hearthstonejson/compare/0.1.0...0.1.1